c-alpha atom support

+++ MathTools/Vector3D.h
@@ -11,11 +11,11 @@ namespace wustl_mm {
 		typedef Vector3DTemplate<double> Vector3DDouble;
 
 		inline Vector3DFloat Vector3DInt2Float(Vector3DInt & v) {
-			return Vector3DFloat(v.X(), v.Y(), v.Z());
+			return Vector3DFloat((float)v.X(), (float)v.Y(), (float)v.Z());
 		}
 
 		inline Vector3DDouble Vector3DInt2Double(Vector3DInt & v) {
-			return Vector3DDouble(v.X(), v.Y(), v.Z());
+			return Vector3DDouble((double)v.X(), (double)v.Y(), (double)v.Z());
 		}
